Lab 4/src/Extract/requirements.txt
blinker
certifi==2025.4.26
charset-normalizer==3.4.2
click
Flask
idna==3.10
itsdangerous
Jinja2
kafka-python==2.2.11
MarkupSafe
pendulum==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug
